+++ src/models/Author.java
@@ -1,0 +1,17 @@
+package models;
+
+/**
+ * Художник комикса
+ */
+public class Author {
+
+    private String name;
+
+    public Author(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
