+++ src/models/ComicPricingBook.java
@@ -1,0 +1,31 @@
+package models;
+
+/**
+ * Книга покупок и продаж
+ */
+public class ComicPricingBook {
+
+    private double costPrice;
+    private double sellingPrice;
+
+    public ComicPricingBook(double costPrice, double sellingPrice) {
+        this.costPrice = costPrice;
+        this.sellingPrice = sellingPrice;
+    }
+
+    public double getCostPrice() {
+        return costPrice;
+    }
+
+    public void setCostPrice(double costPrice) {
+        this.costPrice = costPrice;
+    }
+
+    public double getSellingPrice() {
+        return sellingPrice;
+    }
+
+    public void setSellingPrice(double sellingPrice) {
+        this.sellingPrice = sellingPrice;
+    }
+}
