+++ src/models/Genre.java
@@ -1,0 +1,17 @@
+package models;
+
+/**
+ * Жанр комикса
+ */
+public class Genre {
+
+    private String name;
+
+    public Genre(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
