+++ src/models/Publishing.java
@@ -1,0 +1,17 @@
+package models;
+
+/**
+ * Издатель
+ */
+public class Publishing {
+
+    private String name;
+
+    public Publishing(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
